+++ App/Controllers/FunctionsController.php
@@ -5,16 +5,22 @@ namespace App\Controllers;
 class FunctionsController extends Controller
 {
     function new () {
-        return $this->view("functions.new", "layout_admin");
+        if($this->isLoggedIn()){
+            return $this->view("functions.new", "layout_admin");
+        }
+        $this->askLogin();
     }
     function update () {
-        $id = htmlspecialchars($_GET['id']);
-        $process = $this->getFunctionProcessor();
-        $data = $process->function->findOne($id)->fetch();
-        return $this->view("functions.update", "layout_admin", ['data' => $data]);
+        if($this->isLoggedIn()){
+            $id = htmlspecialchars($_GET['id']);
+            $process = $this->getFunctionProcessor();
+            $data = $process->function->findOne($id)->fetch();
+            return $this->view("functions.update", "layout_admin", ['data' => $data]);
+        }
+        $this->askLogin();
     }
     function _update () {
-        if($this->isPostMethod()){
+        if($this->isPostMethod() && $this->isLoggedIn()){
             $id = htmlspecialchars($_GET['id']);
             $processor = $this->getFunctionProcessor();
             $processor->createFunctionProcess();
@@ -27,7 +33,7 @@ class FunctionsController extends Controller
         }
     }
     function _new () {
-        if($this->isPostMethod()){
+        if($this->isPostMethod() && $this->isLoggedIn()){
             $processor = $this->getFunctionProcessor();
             $processor->createFunctionProcess();
             if($processor->hasErrors()){
@@ -40,15 +46,21 @@ class FunctionsController extends Controller
     }
     public function all()
     {
-        $func = $this->getFunctionProcessor();
-        $functions = $func->getAll();
-        return $this->view("functions.all", "layout_admin", ['functions' => $functions]);
+        if($this->isLoggedIn()){
+            $func = $this->getFunctionProcessor();
+            $functions = $func->getAll();
+            return $this->view("functions.all", "layout_admin", ['functions' => $functions]);
+        }
+        $this->askLogin();
     }
      public function delete(){
-        $id = htmlspecialchars($_GET['id']);
-        $process = $this->getFunctionProcessor();
-        $process->delete($id);
-        $this->redirect("/admin/functions");
+        if($this->isLoggedIn()){
+            $id = htmlspecialchars($_GET['id']);
+            $process = $this->getFunctionProcessor();
+            $process->delete($id);
+            $this->redirect("/admin/functions");
+        }
+        $this->askLogin();
         //return $this->view("search_domains.all", "layout_admin", ['fsd' => $fsd]);
     }
