+++ App/Controllers/PromotionsController.php
@@ -6,15 +6,17 @@ class PromotionsController extends Controller
 {
     public function new()
     {
-        $dep = $this->getDepartmentProcessor();
-        $departments = $dep->getAll();
-        $fac = $this->getFacultyProcessor();
-        $faculties = $fac->getAll();
-        return $this->view("promotions.new", "layout_admin",['fac' => $faculties, 'dep' => $departments]);
+        if($this->isLoggedIn()){
+            $dep = $this->getDepartmentProcessor();
+            $departments = $dep->getAll();
+            $fac = $this->getFacultyProcessor();
+            $faculties = $fac->getAll();
+            return $this->view("promotions.new", "layout_admin",['fac' => $faculties, 'dep' => $departments]);
+        }
+        $this->askLogin();
     }
     public function _new(){
-        if($this->isPostMethod()){
-
+        if($this->isPostMethod() && $this->isLoggedIn()){
             $dep = $this->getDepartmentProcessor();
             $departments = $dep->getAll();
             $fac = $this->getFacultyProcessor();
@@ -22,7 +24,6 @@ class PromotionsController extends Controller
     
             $prom = $this->getPromotionProcessor();
             $prom->createPromotionProcess();
-    
             if($prom->hasErrors()){
                 return $this->view("promotions.new", "layout_admin", ['fac' => $faculties, 'dep' => $departments, 'errors' => $prom->getErrors()]);
             }else{
@@ -34,28 +35,31 @@ class PromotionsController extends Controller
     }
     public function all()
     {
-        $prom = $this->getPromotionProcessor();
-        $promotions = $prom->getAll();
-        return $this->view("promotions.all", "layout_admin", ['prom' =>$promotions]);
+        if($this->isLoggedIn()){
+            $prom = $this->getPromotionProcessor();
+            $promotions = $prom->getAll();
+            return $this->view("promotions.all", "layout_admin", ['prom' =>$promotions]);
+        }
+        $this->askLogin();
     }
     public function update(){
-        $id = htmlspecialchars($_GET['id']);
-        $dep = $this->getDepartmentProcessor();
-        $departments = $dep->getAll();
-        $fac = $this->getFacultyProcessor();
-        $faculties = $fac->getAll();
-
-        $prom = $this->getPromotionProcessor();
-
-        $result = $prom->promotion->findOne($id);
-        if($result !== false){
-            $data = $prom->loadData($result)[0];
-            return $this->view("promotions.update", "layout_admin", ['fac' => $faculties, 'dep' => $departments, 'data' => $data]);
+        if($this->isLoggedIn()){
+            $id = htmlspecialchars($_GET['id']);
+            $dep = $this->getDepartmentProcessor();
+            $departments = $dep->getAll();
+            $fac = $this->getFacultyProcessor();
+            $faculties = $fac->getAll();
+            $prom = $this->getPromotionProcessor();
+            $result = $prom->promotion->findOne($id);
+            if($result !== false){
+                $data = $prom->loadData($result)[0];
+                return $this->view("promotions.update", "layout_admin", ['fac' => $faculties, 'dep' => $departments, 'data' => $data]);
+            }
         }
-
+        $this->askLogin();
     }
     public function _update(){
-        if ($this->isPostMethod()) {
+        if ($this->isPostMethod() && $this->isLoggedIn()) {
             $id = htmlspecialchars($_GET['id']);
             $dep = $this->getDepartmentProcessor();
             $departments = $dep->getAll();
@@ -75,10 +79,13 @@ class PromotionsController extends Controller
 
     }
      public function delete(){
-        $id = htmlspecialchars($_GET['id']);
-        $process = $this->getPromotionProcessor();
-        $process->delete($id);
-        $this->redirect("/admin/promotions");
+        if($this->isLoggedIn()){
+            $id = htmlspecialchars($_GET['id']);
+            $process = $this->getPromotionProcessor();
+            $process->delete($id);
+            $this->redirect("/admin/promotions");
+        }
+        $this->askLogin();
     }
    
 }
